Move file-size-text.tsx into mod-item
Since FileSizeText will only ever be used on ModItems
